Updated sonification code
Some key necessary classes:
- Sonifier is the thing that anyone using the library needs. There can
only be one of this
- DataSource is a data source. It holds a bunch of variables related
to data streaming from that source and how to process it
- Datum is an individual data point from the data source

The sonification library is now organized into "displays" and
"templates"

Displays include
- DatumDisplay (which just logs to console)
- Sonify (inherits from datum and expects you to provide an AudioNode)
- NoteSonify (creates an audio node oscillator to play a note)
- NoiseSonify (creates an audio node to play noise. Untested. Should
be used with a FilterRangeTemplate)

Templates include:
- Template (base class, sets up key methods, including associating a
display with any tempalte
- ScaleTemplate (shifts a datum from one range to another. Stores
result in Datum.adjustedValue)
- FilterRangeTemplate (decides whether datum should be displayed or
not based on a range)
- NoteTemplate (subclass of ScaleTemplate that assigns a
NoteSonify display to it).

There are a couple of demos. I've only tested one so far, and it's not
100% perfect (the first note seems to play for a very long time before
notes start to change, haven't figured out why yet). The key demo to
look at is DemoSimple.tsx. Here are the most important lines of code:
     	// get the sonifier
        this.sonifierInstance = Sonifier.getSonifierInstance()
	// create a data source
        let source = new DataSource(this.sourceId, "SimpleDemo");
	// specify how it should be displayed
        source.addTemplate(new NoteTemplate());
	// add the source to the sonifier instance
        this.sonifierInstance.addSource(this.sourceId, source);
	// store the data set's range in source
        source.addCalculator("max", (datum: Datum, stat: number) => Math.max(...dummyData),0);
        source.addCalculator("min", (datum: Datum, stat: number) =>
	Math.min(...dummyData),0);

It's not all in one place because the demo class also attempts to
schedule the data slowly to simulate a stream but you get the idea...