Improved scale template
ScaleTemplate now takes a function that it uses for the conversion
between domain and range. This defaults to linear

NoteTemplate uses this to improve the quality of the audio output. I'm
using a log function right now, and it sounds better, but we need to
investigate the right function more...