Modification to make use of streams
Simplified timed release of data to make use of streams. See
DemoSimple.onPlay() for an example of how this is set up.
PlayDataSlowly is no longer necessary.

This required some other changes as well.
- DataSources now have a
stream (which is an Observable) and a subscription (which is their
subscription to that stream). They also have a stop and a start method
for stopping and starting the display of the stream and resetting
things. Right now I am assuming data sources only have one stream.
DataSources also have a handleEndStream. This is SUPPOSED to be called
when the stream ends automatically but I haven't been able to get that
working. However I do call it directly at apporpriate times (like when
the data set is changed out) and it works properly when it is called.
- Templates now have a link to their source and a stop and start
method as well.
- Displays have a stop and start method. For example, note sonify
replaces its oscillator when stop is called.