Not fully debugged, but shifting to a stream based model
Streams are used for two different things
1) Changes to OutputState as defined in OutputStateChange Enum
2) Processing data as it arrives

From an end user perspective, this means three things (see how the demos
are updated).

- Call OutputEngine.getInstance().subscribe([handler]) to find out
about changes in output state. Handler receives an
OutputStateChange.[type] variable and decides what to do
- Share information about state changes by calling
OutputEngine.getInstance().next(OutputStateChange.[type])

Intrenally to the toolkit, this also means a lot of things have
changed
- Most classes are now extensions of BehaviorSubject<StateDatum, Datum
or number> depending on what level of information the class needs.
StateDatum is just a dictionary holding both a Datum and an
OutputState. BehaviorSubjects can both receive and stream data
- Most classes no longer need to keep track of whether data is
streaming, output state, and so on.

*OutputEngine*
- still keeps a list of its data sinks (not sure if this is needed,
but it allows us to assert that all sinks have unique identifiers
which (might?) be useful
- when a sink is added, calls setStream to subscribe the sink to the OutputEngine
- when a sink is deleted, calls complete() on that sink to end its
stream so that all output associated with it stops
- has legacy support for pushPoint, which just calls sink.next() with
the data. Not sure if this actually works.
- when setStream is called, creates a pipe that does not send repeated
copies of OutputStateChange events (i.e. no PLAY followed by PLAY),
and combines OutputStateChanges with the data stream and then pipes
the combined stream into the associated sink

*DataSink* is much simpler than it used to be, because it no longer
- whenever a dataHandler is added, that dataHandler's
setupSubscription method is called, passing the DataSink in, so the
dataHandler can add filtering & so on before subscribing itself to the
sink
-

*DataHandler* now just subscribes an output to itself with the
following filters
- only sends data (not state)
- filters data out when the state is not Play

It also handles its own subscription to its sink, just so that more
sophisticated handlers can override that for example to filter.

Finally, because we assume outputs don't want state bundled with
datum (and won't pass the bundle on), we call a separate method when
there is a state update called output.setOutputState(state) when an
output state change arrives.

This makes subclasses very simple. For example
- FilterRangeHandler has a domain that is allowable, and just
overrides setupSubscription with a .pipe(filter() => insideDomain())
check
- ScaleHandler has a domain and a range and a conversion function. It overrides
setupSubscprtion to apply the conversion function using .pipe(map()=>...
- NoteHadler provides scaleHandler with a function that converts from
a domain to a range that is audible using mel conversions, and a
NoteSonify output, in its
constructor and needs to do nothing else

*DatumOutput**
does almost nothing now. It is subscribed by its handler to data and
assumes it will only get data when it should be displayed. it
just prints it to console.out.

Instead of *update* all outputs now implement *next*
In addition, since they are subscribed to Datum, they have actual
functions for start(), stop() and pause(). Most of these classes are
pretty much unchanged from their pre-stream implementations.

*Statistics*
All statistics, even static ones, are implemeted as streams now.
- Statistics They filter out (1) any unspecified data (can happen
because of our joint stream) and (2) assume that they don't get
anything they shouln't record
- I've implemented some sample statistics but not tested them much.